fix looping behaviour

+++ app/jobs/reconcile_rating_topics_job.rb
@@ -6,6 +6,7 @@ class ReconcileRatingTopicsJob < ApplicationJob
     topics = rating.topic_names.map { |t| Topic.find_or_create_for_name(t) }
     topics += active_citation_topics(rating)
     topic_ids = topics.map(&:id).sort # need to sort for comparison
+    return if topics_match_citation_topics?(rating, topic_ids)
     rating.rating_topics.where.not(topic_id: topic_ids).destroy_all
     (topic_ids - rating.rating_topics.pluck(:topic_id)).each do |i|
       RatingTopic.create(rating_id: rating.id, topic_id: i)
@@ -54,7 +55,7 @@ class ReconcileRatingTopicsJob < ApplicationJob
   end
 
   def topics_match_citation_topics?(rating, target_topic_ids = nil)
-    target_topic_ids ||= rating.citation.active_citation_topics.pluck(:id).sort
+    target_topic_ids ||= rating.citation.topics_active.pluck(:id).sort
     target_topic_ids == rating.topics.reorder(:id).pluck(:id)
   end
 end
